requirements.txt
numpy
pandas
scikit-learn
plotly
matplotlib
seaborn
xgboost
wrds
PyYAML